requirements.txt
alembic==1.11.1
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.6.3
tzdata==2023.3
Werkzeug==2.3.6
